fix: 재시작여부, 경기시작여부는 PreparedCarGame에서 담당하도록 수정
